Improve recursive schema
