+++ packages/dds/tree/src/simple-tree/index.ts
@@ -34,7 +34,6 @@ export { SchemaFactory, type ScopedSchemaName } from "./schemaFactory.js";
 export { nodeApi as Tree, TreeApi, TreeNodeEvents } from "./treeApi.js";
 export { toFlexConfig } from "./toFlexSchema.js";
 export {
-	SchemaFactoryRecursive,
 	ObjectFromSchemaRecordUnsafe,
 	TreeFieldFromImplicitFieldUnsafe,
 	TreeNodeFromImplicitAllowedTypesUnsafe,
@@ -47,7 +46,8 @@ export {
 	InsertableTypedNodeUnsafe,
 	NodeBuilderDataUnsafe,
 	NodeFromSchemaUnsafe,
-} from "./schemaFactoryRecursive.js";
+} from "./typesUnsafe.js";
+export { SchemaFactoryRecursive, ValidateRecursiveSchema } from "./schemaFactoryRecursive.js";
 export { getProxyForField } from "./proxies.js";
 
 export {
